+++ pkg/analysis_server/lib/src/cider/local_library_contributor.dart
@@ -193,7 +193,7 @@ class LibraryElementSuggestionBuilder
       return;
     }
 
-    for (var constructor in element.constructors2) {
+    for (var constructor in element.constructors) {
       if (constructor.isPrivate) {
         continue;
       }