+++ pkg/analyzer/lib/src/summary2/constructor_initializer_resolver.dart
@@ -26,7 +26,7 @@ class ConstructorInitializerResolver {
     ];
 
     for (var interfaceElement in interfaceElements) {
-      for (var constructorElement in interfaceElement.constructors2) {
+      for (var constructorElement in interfaceElement.constructors) {
         _constructorElement(interfaceElement, constructorElement);
       }
     }