+++ pkg/analyzer/test/src/dart/resolution/constant_test.dart
@@ -39,7 +39,7 @@ const a = const A();
     );
 
     var aLib = findElement2.import('package:test/a.dart').importedLibrary2!;
-    var aConstructor = aLib.getClass2('A')!.constructors2.single;
+    var aConstructor = aLib.getClass2('A')!.constructors.single;
     var p = aConstructor.formalParameters.single;
     var pf = p.firstFragment as DefaultParameterFragmentImpl;